requirements.txt
camelot-py
matplotlib
nltk
numpy
pandas
wordcloud